Fix notification getNotificationByUUID v1
