+++ src/main/java/com/binar/pemesanantiketpesawat/controller/NotificationController.java
@@ -37,7 +37,9 @@ public class NotificationController {
 
     @GetMapping
     private List<NotificationMessage> getNotificationByUUID(@RequestParam UUID uuidRequest) {
-        log.info("Notification sent successfully");
-        return firebaseMessagingService.getByUUID(uuidRequest);
+        log.info("Received request to get notifications by UUID: '{}'", uuidRequest);
+        List<NotificationMessage> notifications = firebaseMessagingService.getByUUID(uuidRequest);
+        log.info("Retrieved {} notification(s) for UUID: '{}'", notifications.size(), uuidRequest);
+        return notifications;
     }
 }
